solver-service/requirements.txt
Flask==2.2.0
flask-cors
psycopg2-binary==2.9.1
werkzeug==2.2.0
gunicorn==20.1.0
ortools==9.10.4067
python-dotenv
psutil
docker